the initial idea
One is called Bragi: he is renowned for wisdom, and most of all for fluency of speech and skill with words. He knows most of skaldship, and after him skaldship is called bragr, and from his name that one is called bragr-man or -woman, who possesses eloquence surpassing others, of women or of men. His wife is Iðunn.